+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/rqes/RqesConstants.kt
@@ -1,9 +1,5 @@
 package at.asitplus.dif.rqes
 
 object RqesConstants {
-
     const val SCOPE = "credential"
-
-    const val SIGNATURE_QUALIFIER = "eu_eidas_qes"
-
 }