+++ dif-data-classes/src/commonMain/kotlin/at/asitplus/dif/rqes/collection_entries/DocumentLocation.kt
@@ -1,14 +1,17 @@
-package at.asitplus.dif.rqes
+package at.asitplus.dif.rqes.collection_entries
 
+import at.asitplus.dif.rqes.Method
 import io.ktor.http.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
+/**
+ * Class used as part of [AuthorizationDetails]
+ */
 @Serializable
-data class DocumentLocationEntry(
+data class DocumentLocation(
     @SerialName("uri")
-    @Serializable(UrlSerializer::class)
-    val uri: Url,
+    val uri: String,
     @SerialName("method")
     val method: Method,
 )